pyproject.toml
[project]
name = "themecooker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "numpy>=2.3.4",
    "pathlib>=1.0.1",
    "pyyaml>=6.0.3",
    "ruff>=0.14.1",
]
